App.jsxのロジックをcustom hookに分割
- useHabitsStorage: habits/records状態とlocalStorage永続化
- useTheme: テーマ読み込み・保存・反映
- usePullToRefresh: pull-to-refreshジェスチャーとスクロール状態

動作は既存と同一。

closes #18
